+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Versioning](http://semver.org/spec/v2.0.0.html) except to the first release.
 
 ### Added
 
+* `Storage` discoverer allows to fetch instance configuration from a cluster
+  configuration in storage using `go-storage`.
+
 ### Changed
 
 ### Fixed